Increased size of ORGretina4M trapezoidal discriminator threshold to match the hardware.
git-svn-id: svn://orca.physics.unc.edu/Orca@9872 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/Gretina4M/ORGretina4MModel.h
@@ -216,7 +216,7 @@ enum Gretina4MFIFOStates {
     BOOL			easySelected[kNumGretina4MChannels];
     
     int				ledThreshold[kNumGretina4MChannels];
-    short			trapThreshold[kNumGretina4MChannels];
+    long			trapThreshold[kNumGretina4MChannels];
     short           poleZeroMult[kNumGretina4MChannels];
     short			downSample;
     short			mrpsrt[kNumGretina4MChannels];
